# docs/api/openapi-mvp3-draft.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ontology Platform Backend",
    "description": "Ontology Platform MVP backend API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/publish-candidates": {
      "get": {
        "tags": [
          "Publish"
        ],
        "summary": "List candidate publish eligibility",
        "operationId": "list_publish_candidates_api_v1_projects__project_id__publish_candidates_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "candidate_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CandidateKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublishEligibility"
                  },
                  "title": "Response List Publish Candidates Api V1 Projects  Project Id  Publish Candidates Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/publish-jobs": {
      "post": {
        "tags": [
          "Publish"
        ],
        "summary": "Create publish job",
        "operationId": "create_publish_job_api_v1_projects__project_id__publish_jobs_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishJobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJob"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      },
      "get": {
        "tags": [
          "Publish"
        ],
        "summary": "List publish jobs",
        "operationId": "list_publish_jobs_api_v1_projects__project_id__publish_jobs_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublishJob"
                  },
                  "title": "Response List Publish Jobs Api V1 Projects  Project Id  Publish Jobs Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publish-jobs/{publish_job_id}": {
      "get": {
        "tags": [
          "Publish"
        ],
        "summary": "Get publish job",
        "operationId": "get_publish_job_api_v1_publish_jobs__publish_job_id__get",
        "parameters": [
          {
            "name": "publish_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Publish Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJob"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publish-jobs/{publish_job_id}/run": {
      "post": {
        "tags": [
          "Publish"
        ],
        "summary": "Run pending publish job",
        "operationId": "run_publish_job_endpoint_api_v1_publish_jobs__publish_job_id__run_post",
        "parameters": [
          {
            "name": "publish_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Publish Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJob"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/published-graph/versions": {
      "get": {
        "tags": [
          "Publish"
        ],
        "summary": "List published graph versions",
        "operationId": "list_published_graph_versions_api_v1_projects__project_id__published_graph_versions_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublishedGraphVersion"
                  },
                  "title": "Response List Published Graph Versions Api V1 Projects  Project Id  Published Graph Versions Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/published-graph/current": {
      "get": {
        "tags": [
          "Publish"
        ],
        "summary": "Get current published graph snapshot",
        "operationId": "get_current_published_graph_api_v1_projects__project_id__published_graph_current_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedGraphSnapshot"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/published-graph/versions/{version_id}": {
      "get": {
        "tags": [
          "Publish"
        ],
        "summary": "Get published graph version",
        "operationId": "get_published_graph_version_api_v1_published_graph_versions__version_id__get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedGraphVersion"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/published-graph/versions/{version_id}/entities": {
      "get": {
        "tags": [
          "Publish"
        ],
        "summary": "List published entities",
        "operationId": "list_published_entities_api_v1_published_graph_versions__version_id__entities_get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          },
          {
            "name": "class_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublishedEntity"
                  },
                  "title": "Response List Published Entities Api V1 Published Graph Versions  Version Id  Entities Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/published-graph/versions/{version_id}/relations": {
      "get": {
        "tags": [
          "Publish"
        ],
        "summary": "List published relations",
        "operationId": "list_published_relations_api_v1_published_graph_versions__version_id__relations_get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          },
          {
            "name": "relation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Relation Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublishedRelation"
                  },
                  "title": "Response List Published Relations Api V1 Published Graph Versions  Version Id  Relations Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
...
      "CandidateKind": {
        "type": "string",
        "enum": [
          "ENTITY",
          "RELATION",
          "PROPERTY_VALUE"
        ],
        "title": "CandidateKind"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublishEligibility": {
        "properties": {
          "candidate_kind": {
            "$ref": "#/components/schemas/CandidateKind"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/PublishEligibilityReasonCode"
            },
            "type": "array",
            "title": "Reasons"
          },
          "review_status": {
            "$ref": "#/components/schemas/CandidateReviewStatus"
          },
          "publish_status": {
            "$ref": "#/components/schemas/PublishStatus"
          },
          "validation_status": {
            "$ref": "#/components/schemas/ValidationStatus"
          },
          "has_evidence": {
            "type": "boolean",
            "title": "Has Evidence"
          },
          "has_warning_reason": {
            "type": "boolean",
            "title": "Has Warning Reason"
          }
        },
        "type": "object",
        "required": [
          "candidate_kind",
          "candidate_id",
          "eligible",
          "reasons",
          "review_status",
          "publish_status",
          "validation_status",
          "has_evidence",
          "has_warning_reason"
        ],
        "title": "PublishEligibility"
      },
...
      "PublishJobCreateRequest": {
        "properties": {
          "ontology_version_id": {
            "type": "string",
            "title": "Ontology Version Id"
          },
          "candidate_refs": {
            "items": {
              "$ref": "#/components/schemas/CandidateRef"
            },
            "type": "array",
            "title": "Candidate Refs"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "ontology_version_id",
          "candidate_refs"
        ],
        "title": "PublishJobCreateRequest"
      },
...
      "PublishedGraphSnapshot": {
        "properties": {
          "version": {
            "$ref": "#/components/schemas/PublishedGraphVersion"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/PublishedEntity"
            },
            "type": "array",
            "title": "Entities"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/PublishedRelation"
            },
            "type": "array",
            "title": "Relations"
          }
        },
        "type": "object",
        "required": [
          "version",
          "entities",
          "relations"
        ],
        "title": "PublishedGraphSnapshot"
      },
      "PublishedGraphVersion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "ontology_version_id": {
            "type": "string",
            "title": "Ontology Version Id"
          },
          "publish_job_id": {
            "type": "string",
            "title": "Publish Job Id"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "summary": {
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "version",
          "ontology_version_id",
          "publish_job_id",
          "is_current",
          "created_at"
        ],
        "title": "PublishedGraphVersion"
      },
...
    }
  }
}